# [Perf] Vectorize static / dynamic INT8 quant kernels