--- a/csrc/quantization/compressed_tensors/int8_quant_kernels.cu
+++ b/csrc/quantization/compressed_tensors/int8_quant_kernels.cu
@@ -3,6 +3,7 @@
 #include <cmath>
 
 #include "../../dispatch_utils.h"
+#include "../vectorization.cuh"
 
 #ifndef USE_ROCM
   #include <cub/util_type.cuh>
@@ -107,16 +108,39 @@ template <typename scalar_t, typename scale_type>
 __global__ void static_scaled_int8_quant_kernel(
     scalar_t const* __restrict__ input, int8_t* __restrict__ out,
     scale_type const* scale_ptr, const int hidden_size) {
+  int const VEC_SIZE = 4;
   int const tid = threadIdx.x;
   int64_t const token_idx = blockIdx.x;
-  scale_type const scale = *scale_ptr;
+  scale_type const scale_val = *scale_ptr;
 
   // Must be performed using 64-bit math to avoid integer overflow.
   out += token_idx * hidden_size;
   input += token_idx * hidden_size;
 
-  for (int i = tid; i < hidden_size; i += blockDim.x) {
-    out[i] = float_to_int8_rn(static_cast<float>(input[i]) / scale);
+  vec4_t<scalar_t> const* vec_input =
+      reinterpret_cast<vec4_t<scalar_t> const*>(input);
+  q8x4_t<int8_t>* vec_out = reinterpret_cast<q8x4_t<int8_t>*>(out);
+  int const num_vec_elements = hidden_size / VEC_SIZE;
+  for (int i = tid; i < num_vec_elements; i += blockDim.x) {
+    vec4_t<scalar_t> in_vec = vec_input[i];
+    q8x4_t<int8_t> out_vec;
+    out_vec.val[0] =
+        float_to_int8_rn(static_cast<float>(in_vec.val[0]) / scale_val);
+    out_vec.val[1] =
+        float_to_int8_rn(static_cast<float>(in_vec.val[1]) / scale_val);
+    out_vec.val[2] =
+        float_to_int8_rn(static_cast<float>(in_vec.val[2]) / scale_val);
+    out_vec.val[3] =
+        float_to_int8_rn(static_cast<float>(in_vec.val[3]) / scale_val);
+
+    vec_out[i] = out_vec;
+  }
+  // handle the rest
+  int base = num_vec_elements * VEC_SIZE + tid;
+  int rem = hidden_size % VEC_SIZE;
+  for (int i = tid; i < rem; i += blockDim.x) {
+    out[base + i] =
+        float_to_int8_rn(static_cast<float>(input[base + i]) / scale_val);
   }
 }