--- a/csrc/quantization/compressed_tensors/int8_quant_kernels.cu
+++ b/csrc/quantization/compressed_tensors/int8_quant_kernels.cu
@@ -116,10 +116,11 @@ __global__ void static_scaled_int8_quant_kernel(
   const scalar_t* row_in = input + blockIdx.x * hidden_size;
   int8_t* row_out = output + blockIdx.x * hidden_size;
 
-  vectorize_with_alignment(row_in, row_out, hidden_size, tid, stride,
-                           [=] __device__(int8_t& dst, const scalar_t& src) {
-                             dst = float_to_int8_rn(float(src) / scale);
-                           });
+  vectorize_with_alignment<16>(
+      row_in, row_out, hidden_size, tid, stride,
+      [=] __device__(int8_t& dst, const scalar_t& src) {
+        dst = float_to_int8_rn(float(src) / scale);
+      });
 }
 
 template <typename scalar_t, typename scale_t, typename azp_t>
@@ -135,11 +136,12 @@ __global__ void static_scaled_int8_azp_quant_kernel(
   const scalar_t* row_in = input + blockIdx.x * hidden_size;
   int8_t* row_out = output + blockIdx.x * hidden_size;
 
-  vectorize_with_alignment(row_in, row_out, hidden_size, tid, stride,
-                           [=] __device__(int8_t& dst, const scalar_t& src) {
-                             float v = float(src) * inv_s;
-                             dst = int32_to_int8(float_to_int32_rn(v) + azp);
-                           });
+  vectorize_with_alignment<16>(
+      row_in, row_out, hidden_size, tid, stride,
+      [=] __device__(int8_t& dst, const scalar_t& src) {
+        float v = float(src) * inv_s;
+        dst = int32_to_int8(float_to_int32_rn(v) + azp);
+      });
 }
 
 template <typename scalar_t, typename scale_t>
@@ -171,10 +173,11 @@ __global__ void dynamic_scaled_int8_quant_kernel(
   float inv_s = (absmax == 0.f) ? 0.f : 127.f / absmax;
 
   // 2. quantize
-  vectorize_with_alignment(row_in, row_out, hidden_size, tid, stride,
-                           [=] __device__(int8_t& dst, const scalar_t& src) {
-                             dst = float_to_int8_rn(float(src) * inv_s);
-                           });
+  vectorize_with_alignment<16>(
+      row_in, row_out, hidden_size, tid, stride,
+      [=] __device__(int8_t& dst, const scalar_t& src) {
+        dst = float_to_int8_rn(float(src) * inv_s);
+      });
 }
 
 // MinMax structure to hold min and max values in one go
@@ -252,11 +255,12 @@ __global__ void dynamic_scaled_int8_azp_quant_kernel(
   azp_t azp = azp_sh;
 
   // 2. quantize
-  vectorize_with_alignment(row_in, row_out, hidden_size, tid, stride,
-                           [=] __device__(int8_t& dst, const scalar_t& src) {
-                             float v = float(src) * inv_s;
-                             dst = int32_to_int8(float_to_int32_rn(v) + azp);
-                           });
+  vectorize_with_alignment<16>(
+      row_in, row_out, hidden_size, tid, stride,
+      [=] __device__(int8_t& dst, const scalar_t& src) {
+        float v = float(src) * inv_s;
+        dst = int32_to_int8(float_to_int32_rn(v) + azp);
+      });
 }
 
 }  // namespace vllm