--- a/csrc/quantization/vectorization_utils.cuh
+++ b/csrc/quantization/vectorization_utils.cuh
@@ -1,4 +1,3 @@
-// vector_utils.cuh
 #pragma once
 #include "vectorization.cuh"