## Essential Elements of an Effective PR Description Checklist
- [x] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [x] The test plan, such as providing test command.
- [x] The test results, such as pasting the results comparison before and after, or e2e results

## Motivation

The `scaled_int8_quant` kernel family still relied on scalar loads/stores, leaving significant throughput untapped. Issue #18866 highlighted vectorization as the next major opportunity. Thank @mgoin for the great issue!

And thank @ztang2370 for the good start https://github.com/vllm-project/vllm/pull/19062, https://github.com/vllm-project/vllm/pull/19109 by @mgoin for the great incremental compilation document.

## What’s in this PR

1. **Kernel vectorization (VEC_SIZE = 16)**
   * `static_scaled_int8_quant_kernel`
   * `static_scaled_int8_azp_quant_kernel`
   * `dynamic_scaled_int8_quant_kernel`
   * `dynamic_scaled_int8_azp_quant_kernel`

Make a abstraction in `csrc/quantization/vectorization_utils.cuh` so that this func can be reused later.

2. **Zero-point refactor**
   * Merges the min/max scan with the quantization loop to cut one global-memory pass.

3. **Benchmark tooling**
   * Adds `benchmarks/kernels/bench_int8_gemm.py` for side-by-side BF16 vs. INT8 micro-benchmarks.


## Test

Tested on H100:

**Accuracy validation**: `lm_eval --model vllm --model_args pretrained=RedHatAI/Meta-Llama-3.1-8B-Instruct-quantized.w8a8,max_model_len=32768 --trust_remote_code --tasks gsm8k --num_fewshot 5 --batch_size auto`

```bash
before:

|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.7733|±  |0.0115|
|     |       |strict-match    |     5|exact_match|↑  |0.7604|±  |0.0118|

after:

|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.7733|±  |0.0115|
|     |       |strict-match    |     5|exact_match|↑  |0.7604|±  |0.0118|
```

**End to end throughput**: `vllm bench throughput --model RedHatAI/Meta-Llama-3.1-8B-Instruct-quantized.w8a8 --load-format dummy --input-len 1000 --output-len 100 --max-model-len 32768`

```bash
before:

Throughput: 35.31 requests/s, 38822.03 total tokens/s, 3531.41 output tokens/s
Throughput: 34.61 requests/s, 38026.74 total tokens/s, 3461.44 output tokens/s

after:

Throughput: 37.93 requests/s, 41685.29 total tokens/s, 3792.61 output tokens/s
Throughput: 38.57 requests/s, 42403.11 total tokens/s, 3856.95 output tokens/s
```

**Kernel flops test** with `bench_int8_gemm.py`

Before:

| batch_size | torch-bf16 | int8-tensor-w-tensor-a | int8-channel-w-token-a | int8-tensor-w-tensor-a-noquant | int8-channel-w-token-a-noquant |
|---:|---:|---:|---:|---:|---:|
| 1 | 3.03654 | 3.92181 | 3.60902 | 5.25341 | 5.21451 |
| 16 | 57.3562 | 63.2098 | 59.7113 | 86.128 | 87.9333 |
| 64 | 234.556 | 253.902 | 238.535 | 352.491 | 356.684 |
| 128 | 285.766 | 384.991 | 367.512 | 487.873 | 496.294 |
| 256 | 462.462 | 531.309 | 500.083 | 639.285 | 648.289 |
| 512 | 702.442 | 865.719 | 790.126 | 1251.02 | 1253.05 |
| 1024 | 720.504 | 903.559 | 814.465 | 1280.2 | 1322.11 |
| 2048 | 689.227 | 936.958 | 849.424 | 1318.31 | 1303.01 |
| 4096 | 723.089 | 965.33 | 883.938 | 1527.99 | 1404.14 |
| 8192 | 742.027 | 949.993 | 872.24 | 1470.08 | 1351.67 |
| 16384 | 701.668 | 953.361 | 877.533 | 1433.51 | 1332.62 |

After:

| batch_size | torch-bf16 | int8-tensor-w-tensor-a | int8-channel-w-token-a | int8-tensor-w-tensor-a-noquant | int8-channel-w-token-a-noquant |
|------------|-------------|--------------------------|--------------------------|----------------------------------|------------------------------------|
| 1.00       | 3.03        | 3.87                     | 2.98                     | 5.29                             | 5.23                               |
| 16.00      | 57.55       | 62.83                    | 48.69                    | 86.51                            | 88.29                              |
| 64.00      | 226.81      | 250.57                   | 192.12                   | 354.63                           | 358.78                             |
| 128.00     | 283.94      | 383.05                   | 313.06                   | 499.55                           | 503.99                             |
| 256.00     | 463.65      | 525.42                   | 450.55                   | 641.67                           | 647.82                             |
| 512.00     | 672.04      | 934.00                   | 769.39                   | 1265.80                          | 1266.29                            |
| 1024.00    | 684.19      | 1001.14                  | 841.92                   | 1321.50                          | 1301.72                            |
| 2048.00    | 707.74      | 1071.18                  | 873.07                   | 1311.56                          | 1296.30                            |
| 4096.00    | 728.31      | 1111.62                  | 913.55                   | 1530.56                          | 1468.94                            |
| 8192.00    | 743.42      | 1094.42                  | 902.95                   | 1475.45                          | 1432.45                            |
| 16384.00   | 720.14      | 1102.09                  | 917.71                   | 1481.84                          | 1440.72                            |

Unit test:

```bash
pytest test_int8_quant.py 
============================== test session starts ==============================
platform linux -- Python 3.12.11, pytest-8.4.0, pluggy-1.6.0
rootdir: /home/wentao/vllm-source
configfile: pyproject.toml
plugins: anyio-4.9.0
collected 1346 items                                                            

test_int8_quant.py ...................................................... [  4%]
......................................................................... [  9%]
......................................................................... [ 14%]
......................................................................... [ 20%]
......................................................................... [ 25%]
......................................................................... [ 31%]
......................................................................... [ 36%]
......................................................................... [ 41%]
......................................................................... [ 47%]
......................................................................... [ 52%]
......................................................................... [ 58%]
......................................................................... [ 63%]
......................................................................... [ 69%]
......................................................................... [ 74%]
......................................................................... [ 79%]
......................................................................... [ 85%]
......................................................................... [ 90%]
......................................................................... [ 96%]
...................................................                       [100%]

======================= 1346 passed in 252.81s (0:04:12) ========================
```