@yewentao256 2025-06-06
> Nice work overall! I'd like to see the PR title and description to have more context on what exactly the changes are to achieve this "optimization", and what hardware it was tested on to give context to the performance numbers.
> 
> I also would like to see an expansion to the test cases in `vllm/tests/kernels/quantization/test_int8_quant.py` to included these vectorization test cases from `test_fp8_quant.py`
> 
> https://github.com/vllm-project/vllm/blob/8267f9916f9b4ceace3a8daa908d52824068862b/tests/kernels/quantization/test_fp8_quant.py#L17